root/app/requirements.txt
Flask==3.1.0
gunicorn==23.0.0
PyGithub==2.5.0
PyYAML==6.0.2
